requirements.txt
langchain_google_genai
streamlit
langchain
langchain-community
pymupdf
python-dotenv